add dynamoDB query command code
